Add ComposerScripts.php to upstream configuration (#110)
The script facilitates the ability to add a new requirement to the composer.json file located in a site's upstream-configuration
directory. This would only be used internally by our team. We have already instructed users in
https://drupalkit.its.utexas.edu/docs/getting_started/pantheon_setup.html#anatomy-of-the-pantheon-codebase that they should not
edit this directory, so the addition of the script falls under that.

The main motivation for adding this is not for direct use, but to match as closely as possible the contents of
pantheon-upstreams/drupal-composer-managed