78 - Set PHP 8 as the default version for sites (#116)
* Set default PHP version in upstream to 8.1

* Do not specify an exact PHP platform requirement.

The 3.11.0 release of the Drupal Kit will update the default PHP
version in the upstream to 8.1 in order to prepare for Drupal 10
in the 3.12.0 release. However, sites can defer this change until
the update to 3.12.0 by setting PHP 7.4 in the pantheon.yml file.
To accommodate both PHP versions, we remove the platform requirement.
Composer commands run on Pantheon (e.g., composer update) will use
the correct PHP version, and developers running it locally are
responsible for doing the same.